project5 - Hieroglyph image recognition with neural nets/heroku/requirements.txt
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
dill==0.2.5
Flask==0.11.1
Flask-Cors==2.1.2
gunicorn==19.7.1
jsonschema==2.5.1
pickleshare==0.7.2
requests==2.10.0
numpy==1.12.1
pandas==0.20.1
pytz==2017.2
PyWavelets==0.5.2
decorator==4.0.11
olefile==0.44
Pillow==4.1.1
pyparsing==2.2.0
scipy==0.19.0
virtualenv==15.1.0
Werkzeug==0.11.10
python-resize-image==1.1.11
